ssis_gui.py
EDIT: 
Fixed minor issues with the interface. Specifically, the width of the first column, one that is labelled "#", due to its width initially expands as values are being clicked in the combobox causing the size of the rest of the columns to its right to shift and shrink. I added the method 'adjustColumnWidths' to ensure that the width of the first column will retain its size.